+++ example.lua
@@ -106,10 +106,10 @@ end
 
 mng.xbps_mirror("https://repo-de.voidlinux.org/current")
 mng.package [[
-  xdg-utils fuse mesa-dri man-pages-devel telegram-desktop transmission
+  xdg-utils fuse mesa-dri man-pages-devel telegram-desktop transmission-gtk clang cmake
   zsh git stow curl wget neovim tree-sitter-cli ripgrep eza github-cli love htop tree jq redsocks
   ghostty firefox vlc obs kdenlive audacity ttf-ubuntu-font-family dejavu-fonts-ttf zip unzip awww
-  socklog-void chrony
+  socklog-void chrony fzf qimgv
 ]]
 
 -- use_gnome()